resolved travis tests
