+++ src/test/java/seedu/address/logic/commands/CommandTestUtil.java
@@ -34,8 +34,6 @@ public class CommandTestUtil {
     public static final String VALID_PHONE_BOB = "22222222";
     public static final String VALID_EMAIL_AMY = "amy@example.com";
     public static final String VALID_EMAIL_BOB = "bob@example.com";
-    public static final String VALID_ADDRESS_AMY = "Block 312, Amy Street 1";
-    public static final String VALID_ADDRESS_BOB = "Block 123, Bobby Street 3";
     public static final String VALID_AMOUNT_AMY = "0.00";
     public static final String VALID_AMOUNT_BOB = "0.00";
     public static final String VALID_TAG_HUSBAND = "husband";
